+++ Common/src/main/java/com/tristankechlo/random_mob_sizes/mixin/geckolib/GeoEntityRendererMixin.java
@@ -2,7 +2,7 @@ package com.tristankechlo.random_mob_sizes.mixin.geckolib;
 
 import com.mojang.blaze3d.vertex.PoseStack;
 import com.mojang.blaze3d.vertex.VertexConsumer;
-import com.tristankechlo.random_mob_sizes.mixin_access.MobMixinAddon;
+import com.tristankechlo.random_mob_sizes.mixin_helper.MobMixinAddon;
 import net.minecraft.client.renderer.MultiBufferSource;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.Mob;